Migrate sections controller test
